feat: update subscribe data from hardware

+++ MobileApp/Models/ECGDataPoint.cs
@@ -3,6 +3,6 @@
     public class ECGDataPoint
     {
         public int Time { get; set; }
-        public ushort Value { get; set; }
+        public float Value { get; set; }
     }
 }
